+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added config key `debug`
 - Added module `cogs.debug`
 - Added enum class `const.EmojiStr`
+- Added `lydian-cli` command `logs`
+  - `logs latest`: Returns the most recently modified log file
 
 ### Changed